fix: moving to to a utils file as ES seems to crash the CI
